chore(server): move files to lib
