+++ server/lib/bindFunctions.js
