+++ server/lib/bindFunctions.test.js
